Template information and surface model for D. sechellia reference brain

The DsecI reference brain is a shape averaged template brain generated using the CMTK avg_adm tool. See https://github.com/jefferislab/MakeAverageBrain for the relevant code.

The surface model was constructed in Amira by G. Jefferis and R. Benton using a Z drop correction of the DsecI reference brain in Amira (e^u) followed by a simple threshold (10000), surface simplification to ~ 18,000 faces, default surface smoothing and manual editing in meshlab to remove a small unconnected island of points.

See also

Examples

if (FALSE) {
# Depends on nat
plot3d(DsecI.surf)
plot3d(DsecI)
}