Template information and Surface model of the FCWB FlyCircuit reference brain

The FCWB reference brain is a shape averaged template brain generated using the CMTK avg_adm tool. See https://github.com/jefferislab/MakeAverageBrain for the relevant code.

The FCWB.surf surface model was generated in Amira by constructing an isosurface model based on the FCWB brain.

FCWB

Format

An object of class templatebrain of dimension 1769 x 1026 x 108.

References

Ann-Shyn Chiang, Chih-Yung Lin, Chao-Chun Chuang, Hsiu-Ming Chang, Chang-Huain Hsieh11, Chang-Wei Yeh, Chi-Tin Shih, Jian-Jheng Wu, Guo-Tzau Wang, Yung-Chang Chen, Cheng-Chi Wu, Guan-Yu Chen, Yu-Tai Ching, Ping-Chang Lee, Chih-Yang Lin, Hui-Hao Lin, Chia-Chou Wu, Hao-Wei Hsu, Yun-Ann Huang, Jing-Yi Chen, Hsin-Jung Chiang, Chun-Fang Lu, Ru-Fen Ni, Chao-Yuan Yeh, Jenn-Kang Hwang (2011). Three-dimensional reconstruction of brain-wide wiring networks in Drosophila at single-cell resolution. Current Biology 21, 1-11. doi:10.1016/j.cub.2010.11.056

FCWB is available for download at http://dx.doi.org/10.5281/zenodo.10568

See also

Examples

FCWB
#> === Template Brain === 
#> Name: FlyCircuit Whole Brain 
#> Short Name: FCWB 
#> Type: Average 
#> Sex:  Intersex 
#> Dimensions:1769 x 1026 x 108 voxels
#> Voxel size:
#>   x = 0.318967307692308 microns
#>   y = 0.318427024390244 microns
#>   z = 1 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 563.9342, y = 326.3877, z = 107.
#> Description: An intersex averaged brain formed from 17 female and 9 male brains from the FlyCircuit dataset. 
#> DOI: 10.5281/zenodo.10568
if (FALSE) {
# Depends on nat
plot3d(FCWB.surf)
}