Template information for the Insect Brain Nomenclature reference brain

Template used for the study "A Systematic Nomenclature for the Insect Brain" (doi:10.1016/j.neuron.2013.12.017).

IBN

Format

An object of class templatebrain of dimension 512 x 512 x 135.

Details

Constructed from the green channel of serial sections of a Drosophila melanogaster brain that was tricolor-labeled with a combination of the following reporters:

cytoplasmic DsRed (UAS-DsRed; Verkhusha et al., 2001; red channel),

presynaptic GFP (synaptic-vesicles-targeted UAS-n-syb-GFP; Estes et al., 2000; green channel),

postsynaptic Rdl-hemagglutinin (GABA-receptor-targeted UAS-Rdl-HA; Sanchez-Soriano et al., 2005; blue channel).

These were expressed with the pan-neuronal elav-GAL4 expression driver (C155; Lin and Goodman, 1994).

Note that, because of the excess amount of ectopic protein expression, the presynaptic and postsynaptic labelling (especially the latter) was not completely specific, visualizing not only synaptic sites but also cell bodies and large fiber bundles.

The image was acquired using a confocal laser-scanning microscope LSM510 (Zeiss) with a 40x water-immersion C-Apochromat objective (NA = 1.2). Each section, of 135 total, was recorded with a resolution of 1024 by 1024 pixels and 1.41 micron optical z-slice steps. These were then down-sampled to a resolution of 512 by 512 pixels.

References

Kei Ito, Kazunori Shinomiya, Masayoshi Ito, J. Douglas Armstrong, George Boyan, Volker Hartenstein, Steffen Harzsch, Martin Heisenberg, Uwe Homberg, Arnim Jenett, Haig Keshishian, Linda L. Restifo, Wolfgang Rössler, Julie H. Simpson, Nicholas J. Strausfeld, Roland Strauss, Leslie B. Vosshall, Insect Brain Name Working Group (2013). A systematic nomenclature for the insect brain. Neuron 81 (4), 755-765. doi:10.1016/j.neuron.2013.12.017

Examples

IBN
#> === Template Brain === 
#> Name: Insect Brain Nomenclature 
#> Short Name: IBN 
#> Type: Single brain 
#> Sex:  Female 
#> Dimensions:512 x 512 x 135 voxels
#> Voxel size:
#>   x = 0.64 
#>   y = 0.64 
#>   z = 1.41 
#> Bounding box ():
#>   x = 0, y = 0, z = 0,
#>   x = 327.04, y = 327.04, z = 188.94.
#> Description: Template used for the study "A Systematic Nomenclature for the Insect Brain" (doi: 10.1016/j.neuron.2013.12.01).
#> 
#> Constructed from the green channel of serial sections of a Drosophila melanogaster brain that was tricolor-labeled with a combination of the following reporters:
#>   * cytoplasmic DsRed (UAS-DsRed; Verkhusha et al., 2001; red channel),
#>   * presynaptic GFP (synaptic-vesicles-targeted UAS-n-syb-GFP; Estes et al., 2000; green channel),
#>   * postsynaptic Rdl-hemagglutinin (GABA-receptor-targeted UAS-Rdl-HA; Sanchez-Soriano et al., 2005; blue channel).
#> They are expressed with the pan-neuronal elav-GAL4 expression driver (C155; Lin and Goodman, 1994).
#> 
#> Note that, because of the excess amount of ectopic protein expression, the presynaptic and postsynaptic labelling (especially the latter) was not completely specific, visualizing not only synaptic sites but also cell bodies and large fiber bundles.
#> 
#> The image was acquired using a confocal laser-scanning microscope LSM510 (Zeiss) with a 40x water-immersion C-Apochromat objective (NA = 1.2). Each section, of 135 total, was recorded with a resolution of 1024 by 1024 pixels and 1.41 micron optical z-slice steps. These were then down-sampled to a resolution of 512 by 512 pixels. 
#> DOI: