Template information and surface models for the JRC2018 reference brains

The JRC2018F reference brain is an average template brain constructed from brains labelled with brp-SNAP, dehydrated, and mounted in DPX and imaged at 0.19 x 0.19 x 0.38 microns. The image was downsampled in XY to result in a 0.38 micron isotropic voxel size, which we take to be the standard JRC2018F space.

The JRC2018U reference brain was constructed as for JRC2018F brain but pooling both male and female brains.

The JRC2018M reference brain was constructed as for JRC2018F brain but with 26 male brains.

JRC2018F.surf, JRC2018U.surf were generated in Amira from a 2 micron downsampled, Lanczos filtered 8 bit version of the respective template brain. A surface was then generated with a threshold level of 20 in case of brains this was then smoothed and downsampled. See data-raw/JFRC2018_FEMALE_surface.hx, data-raw/JRC2018_UNISEX_surface.hx for details.

JRC2018F

JRC2018U

JRC2018M

JRC2018F.surf

JRC2018U.surf

Format

An object of class templatebrain of dimension 1652 x 768 x 479.

An object of class templatebrain of dimension 1652 x 773 x 456.

An object of class templatebrain of dimension 1561 x 744 x 546.

An object of class hxsurf (inherits from list) of length 4.

An object of class hxsurf (inherits from list) of length 4.

Details

For the central brain, Bogovic and Saalfeld used 36 female individuals (72 images including left-right flips) for the female template, 26 male individuals (52 image with left-right flips) for the male template, and the union of both for the unisex brain template: 62 individuals (124 images with left-right flips).

Downloaded from https://www.janelia.org/open-science/jrc-2018-brain-templates

References

An unbiased template of the Drosophila brain and ventral nerve cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito, Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi: doi:10.1101/376384

See also

Examples

dim(JRC2018F)
#> [1] 1652  768  479
voxdims(JRC2018F)
#> [1] 0.38 0.38 0.38
boundingbox(JRC2018F)
#>        [,1]   [,2]   [,3]
#> [1,]   0.00   0.00   0.00
#> [2,] 627.38 291.46 181.64
#> attr(,"class")
#> [1] "boundingbox"
JRC2018F
#> === Template Brain === 
#> Name: JRC2018F 
#> Short Name: JRC2018F 
#> Type: Average template brain 
#> Sex:  F 
#> Dimensions:1652 x 768 x 479 voxels
#> Voxel size:
#>   x = 0.38 microns
#>   y = 0.38 microns
#>   z = 0.38 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 627.38, y = 291.46, z = 181.64.
#> Description: Average template brain constructed from brains labelled with brp-SNAP presynaptic marker,
#>     dehydrated, and mounted in DPX and then imaged by confocal microscope. 
#> DOI: https://doi.org/10.1101/376384

if (FALSE) {
library(nat)
plot3d(JRC2018F)
plot3d(JRC2018F.surf)
plot3d(JRC2018F.surf, col='red', alpha=0.3)
}
dim(JRC2018U)
#> [1] 1652  773  456
voxdims(JRC2018U)
#> [1] 0.38 0.38 0.38
boundingbox(JRC2018U)
#>        [,1]   [,2]  [,3]
#> [1,]   0.00   0.00   0.0
#> [2,] 627.38 293.36 172.9
#> attr(,"class")
#> [1] "boundingbox"
JRC2018U
#> === Template Brain === 
#> Name: JRC2018U 
#> Short Name: JRC2018U 
#> Type: Average template brain 
#> Sex:  Intersex 
#> Dimensions:1652 x 773 x 456 voxels
#> Voxel size:
#>   x = 0.38 microns
#>   y = 0.38 microns
#>   z = 0.38 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 627.38, y = 293.36, z = 172.9.
#> Description: Average template brain constructed from brains labelled with brp-SNAP presynaptic marker,dehydrated, and mounted in DPX and then imaged by confocal microscope. There were 36 female individuals (72 images including left-right flips) and 26 male individuals (52 image with left-right flips) for a total of 62 individuals (124 images with left-right flips). 
#> DOI: https://doi.org/10.1101/376384

if (FALSE) {
library(nat)
plot3d(JRC2018U)
plot3d(JRC2018U.surf)
plot3d(JRC2018U.surf, col='red', alpha=0.3)
}