Template information and surface models for JRC2018 VNC templates

The JRCVNC2018F reference VNC was constructed as for JRC2018F brain based on 36 individuals, with a resolution of 0.4 x 0.4 x 0.4 microns.

JRCVNC2018F.surf, was generated in Amira from a 2 micron downsampled, Lanczos filtered version of the respective template brain. A surface was then generated with a threshold level of 2800 and then smoothed and downsampled.

JRCVNC2018F

JRCVNC2018F.surf

Format

An object of class templatebrain of dimension 660 x 1342 x 358.

An object of class hxsurf (inherits from list) of length 4.

References

An unbiased template of the Drosophila brain and ventral nerve cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito, Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi: doi:10.1101/376384

See also

Examples

JRCVNC2018F

#> === Template Brain === 
#> Name: JRCVNC2018F 
#> Short Name: JRCVNC2018F 
#> Type: Average template VNC 
#> Sex:  F 
#> Dimensions:660 x 1342 x 358 voxels
#> Voxel size:
#>   x = 0.4 microns
#>   y = 0.4 microns
#>   z = 0.4 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 263.6, y = 536.4, z = 142.8.
#> Description: Average template VNC constructed from VNC labelled with brp-SNAP presynaptic marker,
#>     dehydrated, and mounted in DPX and then imaged by confocal microscope. 
#> DOI: https://doi.org/10.1101/376384

dim(JRCVNC2018F)

#> [1]  660 1342  358

voxdims(JRCVNC2018F)

#> [1] 0.4 0.4 0.4

boundingbox(JRCVNC2018F)

#>       [,1]  [,2]  [,3]
#> [1,]   0.0   0.0   0.0
#> [2,] 263.6 536.4 142.8
#> attr(,"class")
#> [1] "boundingbox"

if (FALSE) {
library(nat)
plot3d(JRCVNC2018F)
plot3d(JRCVNC2018F.surf)
plot3d(JRCVNC2018F.surf, col='red', alpha=0.3)
}