Template information and surface model for the T1 reference brain

Template information and surface model for the T1 reference brain

The surface model was constructed in Amira in the Jefferis Lab using a simple threshold, followed by a surface simplification.

T1

Format

An object of class templatebrain of dimension 768 x 768 x 165.

References

Jai Y. Yu, Makoto I. Kanai, Ebru Demir, Gregory S. X. E. Jefferis, Barry J. Dickson (2010). Cellular Organization of the Neural Circuit that Drives Drosophila Courtship Behavior. Current Biology 20 (18), 1602-1614. doi:10.1016/j.cub.2010.08.025

Jai Y. Yu, Makoto I. Kanai, Ebru Demir, Gregory S. X. E. Jefferis, Barry J. Dickson (2010). Cellular Organization of the Neural Circuit that Drives Drosophila Courtship Behavior. Current Biology 20 (18), 1602-1614. doi:10.1016/j.cub.2010.08.025

Examples

T1

#> === Template Brain === 
#> Name: T1 
#> Short Name: T1 
#> Type: Average 
#> Sex:  Intersex 
#> Dimensions:768 x 768 x 165 voxels
#> Voxel size:
#>   x = 0.548887614080834 microns
#>   y = 0.548887614080834 microns
#>   z = 1 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 420.9968, y = 420.9968, z = 164.
#> Description: BrainBase 
#> DOI: 10.5281/zenodo.10590
# NOT RUN {
# Depends on nat
plot3d(T1.surf)
# }