Additional flybrain-specific data and tools for use with the NeuroAnatomy

This package provides data, such as template brain information and registrations, along with extra functions to extend nat towards easy use for researchers working with fly data.

Template brains

Information on a number of template brains is provided:

FCWB

IBN

IBNWB

IS2

JFRC2

T1

This information is defined in an object of class templatebrain of the same name as the template brain.

Bridging registrations

A number of registrations for bridging data from one template to another are provided.

IS2->FCWB

JFRC2->FCWB

IBN->INBWB

INBWB->JFRC2

IS2->JFRC2

See xform_brain for examples of how to use these registrations. Additional registrations are available at https://github.com/jefferislab/BridgingRegistrations.

Mirroring registrations

A number of registrations for mirroring data from one side of a template brain to the other are provided.

FCWB

IS2

JFRC2

See mirror_brain for examples of how to use these registrations. Additional registrations are available at https://github.com/jefferislab/MirrorRegistrations.

Additional registrations

See the nat.templatebrains package documentation for details of how to add extra registrations.

See also

nat, nat.templatebrains