Template information and Surface model of the FCWB FlyCircuit reference brain

The FCWB reference brain is a shape averaged template brain generated using the CMTK avg_adm tool. See https://github.com/jefferislab/MakeAverageBrain for the relevant code.

The FCWB.surf surface model was generated in Amira by constructing an isosurface model based on the FCWB brain.

References

Ann-Shyn Chiang, Chih-Yung Lin, Chao-Chun Chuang, Hsiu-Ming Chang, Chang-Huain Hsieh11, Chang-Wei Yeh, Chi-Tin Shih, Jian-Jheng Wu, Guo-Tzau Wang, Yung-Chang Chen, Cheng-Chi Wu, Guan-Yu Chen, Yu-Tai Ching, Ping-Chang Lee, Chih-Yang Lin, Hui-Hao Lin, Chia-Chou Wu, Hao-Wei Hsu, Yun-Ann Huang, Jing-Yi Chen, Hsin-Jung Chiang, Chun-Fang Lu, Ru-Fen Ni, Chao-Yuan Yeh, Jenn-Kang Hwang (2011). Three-dimensional reconstruction of brain-wide wiring networks in Drosophila at single-cell resolution. Current Biology 21, 1-11. http://dx.doi.org/10.1016/j.cub.2010.11.056

FCWB is available for download at http://dx.doi.org/10.5281/zenodo.10568

See also

FCWBNP.surf

Examples

## Not run: ------------------------------------
# # Depends on nat
# plot3d(FCWB.surf)
## ---------------------------------------------