Template information for the Insect Brain Nomenclature Whole Brain reference brain

A synthetic whole brain constructed from the Insect Brain Nomenclature template brain. The original confocal stack was duplicated, mirrored about a vertical axis and then displaced by 392 pixels in \(x\) and then merged with the original, with a linear blend in the overlapping region.

The surface model was constructed in Amira in the Jefferis Lab using a simple median filter, followed by thresholding and a surface simplification.

IBNWB

Format

An object of class templatebrain of dimension 905 x 513 x 136.

References

Kei Ito, Kazunori Shinomiya, Masayoshi Ito, J. Douglas Armstrong, George Boyan, Volker Hartenstein, Steffen Harzsch, Martin Heisenberg, Uwe Homberg, Arnim Jenett, Haig Keshishian, Linda L. Restifo, Wolfgang Rössler, Julie H. Simpson, Nicholas J. Strausfeld, Roland Strauss, Leslie B. Vosshall, Insect Brain Name Working Group (2013). A systematic nomenclature for the insect brain. Neuron 81 (4), 755-765. doi:10.1016/j.neuron.2013.12.017

Kei Ito, Kazunori Shinomiya, Masayoshi Ito, J. Douglas Armstrong, George Boyan, Volker Hartenstein, Steffen Harzsch, Martin Heisenberg, Uwe Homberg, Arnim Jenett, Haig Keshishian, Linda L. Restifo, Wolfgang Rössler, Julie H. Simpson, Nicholas J. Strausfeld, Roland Strauss, Leslie B. Vosshall, Insect Brain Name Working Group (2013). A systematic nomenclature for the insect brain. Neuron 81 (4), 755-765. doi:10.1016/j.neuron.2013.12.017

See also

Examples

IBNWB

#> === Template Brain === 
#> Name: Insect Brain Nomenclature Whole Brain 
#> Short Name: IBNWB 
#> Type: Single brain 
#> Sex:  Female 
#> Dimensions:905 x 513 x 136 voxels
#> Voxel size:
#>   x = 0.64 microns
#>   y = 0.64 microns
#>   z = 1.41 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 578.56, y = 327.68, z = 190.35.
#> Description: A synthetic whole brain constructed from the Insect Brain Nomenclature template brain. The original confocal stack was duplicated, mirrored about a vertical axis and then displaced by 392 pixels in x and then merged with the original, with a linear blend in the overlapping region. 
#> DOI: 10.5281/zenodo.10569