Template information and surface model for the JFRC2 reference brain

The JFRC2 reference brain is a spatially calibrated version of the original FlyLight reference brain (JFRC), which was delivered uncalibrated. This in turn was derived from a single female brain stained with nc82, that was imaged at 1 micron spacing in Z. It was then interpolated in Z to give an isotropic voxel size of 0.622088 microns.

The JFRC2.surf surface model was constructed in Amira in the Jefferis Lab using a simple threshold, followed by a surface simplification.

JFRC2

Format

An object of class templatebrain of dimension 1024 x 512 x 218.

References

Arnim Jenett, Gerald M. Rubin, Teri-T.B. Ngo, David Shepherd, Christine Murphy, Heather Dionne, Barret D. Pfeiffer, Amanda Cavallaro, Donald Hall, Jennifer Jeter, Nirmala Iyer, Dona Fetter, Joanna H. Hausenfluck, Hanchuan Peng, Eric T. Trautman, Robert R. Svirskas, Eugene W. Myers, Zbigniew R. Iwinski, Yoshinori Aso, Gina M. DePasquale, Adrianne Enos, Phuson Hulamm, Shing Chun Benny Lam, Hsing-Hsi Li, Todd R. Laverty, Fuhui Long, Lei Qu, Sean D. Murphy, Konrad Rokicki, Todd Safford, Kshiti Shaw, Julie H. Simpson, Allison Sowell, Susana Tae, Yang Yu, Christopher T. Zugates (2012). A GAL4-Driver Line Resource for Drosophila Neurobiology. Cell Reports 2 (4), 991 - 1001. doi:10.1016/j.celrep.2012.09.011

See also

Examples

JFRC2

#> === Template Brain === 
#> Name: JFRC2 
#> Short Name: JFRC2 
#> Type: Single-brain 
#> Sex:  Female 
#> Dimensions:1024 x 512 x 218 voxels
#> Voxel size:
#>   x = 0.622088 microns
#>   y = 0.622088 microns
#>   z = 0.622088 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 636.396, y = 317.887, z = 134.9931.
#> Description: The JFRC2 reference brain is a refractive-index-corrected version of the original FlyLight reference brain (JFRC). 
#> DOI: 10.5281/zenodo.10567
# NOT RUN {
# Depends on nat
plot3d(JFRC2.surf)
# }