Template information and surface model for the JFRC2013 reference brain

The JFRC2013 reference brain is a single female brain. It is a spatially calibrated version of the brain used in Aso et al. 2014, "The neuronal architecture of the mushroom body provides a logic for associative learning".

JFRC2013.surf was generated in Amira from a 2 micron downsampled and median filtered version of the JFRC2013 template brain. Further cleaning was then carried out in MeshLab to remove non-manifold edges and fill holes.

JFRC2013DS is a downsampled version of the JFRC2013 reference brain, designed for use with images taken on a microscope with 20 x magnification. The downsampling has shrunk the z direction, such that the brain appears flattened when compared with the original.

JFRC2013

JFRC2013DS

Format

An object of class templatebrain of dimension 1450 x 725 x 436.

An object of class templatebrain of dimension 1184 x 592 x 218.

Calibration

The isotropic calibration of 0.38 microns is based on a personal communication from Yoshi Aso on 21st May 2014, where he noted:

It is 0.38um isotropic. Due to dehydration steps with EtOH, tissues are about 20% smaller than when mounted in Vectashield or other glycerol based mounting medium.

No attempt was made to correct for this shrinkage artefact.

References

Aso, Y., Sitaraman, D., Ichinose, T., Kaun, K. R., Vogt, K., Belliart-Guerin, G., Placais, P.-Y., Robie, A. A., Yamagata, N., Schnaitmann, C., Rowell, W. J., Johnston, R. M., Ngo, T.-T. B., Chen, N., Korff, W., Nitabach, M. N., Heberlein, U., Preat, T., Branson, K. M., Tanimoto, H. and Rubin, G. M. (2014b). Mushroom body output neurons encode valence and guide memory-based action selection in Drosophila. Elife 3, e04580. doi:10.7554/eLife.04577

Examples

dim(JFRC2013)

#> [1] 1450  725  436

voxdims(JFRC2013)

#> [1] 0.38 0.38 0.38

boundingbox(JFRC2013)

#>        [,1]   [,2]  [,3]
#> [1,]   0.00   0.00   0.0
#> [2,] 550.62 275.12 165.3
#> attr(,"class")
#> [1] "boundingbox"

JFRC2013

#> === Template Brain === 
#> Name: JFRC2013 
#> Short Name: JFRC2013 
#> Type: Single brain 
#> Sex:  Female 
#> Dimensions:1450 x 725 x 436 voxels
#> Voxel size:
#>   x = 0.38 microns
#>   y = 0.38 microns
#>   z = 0.38 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 550.62, y = 275.12, z = 165.3.
#> Description: The JFRC2013 reference brain is a spatially calibrated version of the brain used in Aso et al. 2014, "The neuronal architecture of the mushroom body provides a logic for associative learning". 
#> DOI: 10.7554/eLife.04577

if (FALSE) {
library(nat)
plot3d(JFRC2013)
plot3d(JFRC2013.surf)
plot3d(JFRC2013.surf, col='red', alpha=0.3)
}
dim(JFRC2013DS)

#> [1] 1184  592  218

voxdims(JFRC2013DS)

#> [1] 0.46 0.46 0.46

boundingbox(JFRC2013DS)

#>        [,1]   [,2]  [,3]
#> [1,]   0.00   0.00  0.00
#> [2,] 544.18 271.86 99.82
#> attr(,"class")
#> [1] "boundingbox"

JFRC2013DS

#> === Template Brain === 
#> Name: JFRC2013DS 
#> Short Name: JFRC2013DS 
#> Type: Single brain 
#> Sex:  Female 
#> Dimensions:1184 x 592 x 218 voxels
#> Voxel size:
#>   x = 0.46 microns
#>   y = 0.46 microns
#>   z = 0.46 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 550.62, y = 275.12, z = 165.3.
#> Description: This is a downsampled version of the JFRC2013 reference brain, designed for use with images taken on a microscope with 20 x magnification. The downsampling has shrunk the z direction, such that the brain appears flattened when compared with the original. 
#> DOI: 10.7554/eLife.04577