Template information and surface models for the JRC2018 reference ventral nerve chords

The JRCVNC2018F reference VNC is an average template VNC constructed from VNCs labelled with brp-SNAP, dehydrated, and mounted in DPX and imaged at 0.19 x 0.19 x 0.38 microns. The image was downsampled to result in a 0.4 micron isotropic voxel size, which we take to be the standard JRCVNC2018F space.

The JRCVNC2018U reference VNC was constructed as for JRCVNC2018F VNC but pooling both male and female brains.

JRCVNC2018F.surf, was generated in Amira from a 2 micron downsampled, Lanczos filtered version of the respective template VNC. A surface was then generated with a threshold level of 2800 and then smoothed and downsampled.

JRCVNC2018U.surf, was generated in Amira from a 2 micron downsampled, Lanczos filtered 8 bit version of the respective template VNC. A surface was then generated with a threshold level of 20 and then smoothed and downsampled.

JRCVNC2018F

JRCVNC2018U

JRCVNC2018F.surf

JRCVNC2018U.surf

Format

An object of class templatebrain of dimension 660 x 1342 x 358.

An object of class templatebrain of dimension 660 x 1290 x 382.

An object of class hxsurf (inherits from list) of length 4.

An object of class hxsurf (inherits from list) of length 4.

Details

For the VNC, Bogovic and Saalfeld used 36 female individuals (72 images including left-right flips) for the female template, 39 male individuals (78 image with left-right flips) for the male template, and the union of both for the unisex brain template: 75 individuals (150 images with left-right flips).

Downloaded from https://www.janelia.org/open-science/jrc-2018-brain-templates

References

An unbiased template of the Drosophila brain and ventral nerve cord John A Bogovic, Hideo Otsuna, Larissa Heinrich, Masayoshi Ito, Jennifer Jeter, Geoffrey W Meissner, Aljoscha Nern, Jennifer Colonell, Oz Malkesman, Kei Ito, Stephan Saalfeld bioRxiv 376384; doi: doi:10.1101/376384

See also

Examples

JRCVNC2018F

#> === Template Brain === 
#> Name: JRCVNC2018F 
#> Short Name: JRCVNC2018F 
#> Type: Average template VNC 
#> Sex:  F 
#> Dimensions:660 x 1342 x 358 voxels
#> Voxel size:
#>   x = 0.4 microns
#>   y = 0.4 microns
#>   z = 0.4 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 263.6, y = 536.4, z = 142.8.
#> Description: Average template VNC constructed from VNC labelled with brp-SNAP presynaptic marker,
#>     dehydrated, and mounted in DPX and then imaged by confocal microscope. 
#> DOI: https://doi.org/10.1101/376384

dim(JRCVNC2018F)

#> [1]  660 1342  358

voxdims(JRCVNC2018F)

#> [1] 0.4 0.4 0.4

boundingbox(JRCVNC2018F)

#>       [,1]  [,2]  [,3]
#> [1,]   0.0   0.0   0.0
#> [2,] 263.6 536.4 142.8
#> attr(,"class")
#> [1] "boundingbox"

if (FALSE) {
library(nat)
plot3d(JRCVNC2018F)
plot3d(JRCVNC2018F.surf)
plot3d(JRCVNC2018F.surf, col='red', alpha=0.3)
}
JRCVNC2018U

#> === Template Brain === 
#> Name: Janelia 2018 average unisex VNC template 
#> Short Name: JRCVNC2018U 
#> Type: average 
#> Sex:  Unisex 
#> Dimensions:660 x 1290 x 382 voxels
#> Voxel size:
#>   x = 0.4 microns
#>   y = 0.4 microns
#>   z = 0.4 microns
#> Bounding box (microns):
#>   x = 0, y = 0, z = 0,
#>   x = 263.6, y = 515.6, z = 152.4.
#> Description: Average template VNC constructed from 75 individuals (36 female, 39 male), VNC labelled with brp-SNAP presynaptic marker, dehydrated, and mounted in DPX and then imaged by confocal microscope. See Bogovic et al 2018 for details. 
#> DOI: https://doi.org/10.1101/376384

dim(JRCVNC2018U)

#> [1]  660 1290  382

voxdims(JRCVNC2018U)

#> [1] 0.4 0.4 0.4

boundingbox(JRCVNC2018U)

#>       [,1]  [,2]  [,3]
#> [1,]   0.0   0.0   0.0
#> [2,] 263.6 515.6 152.4
#> attr(,"class")
#> [1] "boundingbox"

if (FALSE) {
library(nat)
plot3d(JRCVNC2018U)
plot3d(JRCVNC2018U.surf)
plot3d(JRCVNC2018U.surf, col='red', alpha=0.3)
}