Download (or update) and install Jefferis lab registrations

Uses nat.templatebrains::download_reg_repo to download git repositories containing bridging and mirroring registrations for Drosophila melanogaster, simulans and virilis template brains.

download_jefferislab_registrations(repos = c("BridgingRegistrations",
  "MirrorRegistrations", "DrosophilidBridgingRegistrations"))

Arguments

repos
character vector of Jefferis Lab GitHub repositories to download from.

References

For a full description and sample applications of these registrations, please see:

Combining genome-scale Drosophila 3D neuroanatomical data by bridging template brains James D. Manton, Aaron D. Ostrovsky, Lea Goetz, Marta Costa, Torsten Rohlfing, Gregory S. X. E. Jefferis bioRxiv doi: 10.1101/006353

See also

Examples

# NOT RUN {
# to download just one of these repositories do e.g.:
download_jefferislab_registrations("Mirror")

# to download your own registration repositories for use do
download_reg_repo("https://github.com/myuser/myrepo")
# }